Cleaned up timeout code from in memory channel message dispatcher (reusing the one in message receiver) Splitted message sender and message receiver
